Add ExperimentalPanel
